third_party/WebKit/LayoutTests/inspector/console/console-copy-truncated-text-expected.txt

CONSOLE MESSAGE: line 3: www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com
CONSOLE MESSAGE: line 5: www.bar.com
CONSOLE MESSAGE: line 6: www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com
CONSOLE MESSAGE: line 7: www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com www.bar.com www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com
CONSOLE MESSAGE: line 8: www.01234567890123456789zfoobarz01234567890123456789
CONSOLE MESSAGE: line 9: www.01234567890123456789zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz01234567890123456789
CONSOLE MESSAGE: line 10: www.zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz.com
CONSOLE MESSAGE: line 11: www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com
Tests that console copies tree outline messages properly.

Message count: 8
Long url has max length: 150, text: www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z…z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com

Running: testSelectWithinTruncatedUrl

Making selection: 0, 0, 0, 75
Selection length: 75, text: www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z

Making selection: 0, 0, 0, 76
Selection length: 84, text: www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789

Making selection: 0, 0, 0, 150
Selection length: 158, text: www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com

Making selection: 0, 75, 0, 76
Selection length: 9, text: 123456789

Making selection: 0, 75, 0, 150
Selection length: 83, text: 123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com

Making selection: 0, 76, 0, 150
Selection length: 74, text: z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com

Running: testSelectAcrossMultipleMessages

Making selection: 0, 0, 1, 11
Selection length: 177, text: www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com
VM:5 www.bar.com

Making selection: 0, 75, 1, 11
Selection length: 102, text: 123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com
VM:5 www.bar.com

Making selection: 0, 76, 1, 11
Selection length: 93, text: z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com
VM:5 www.bar.com

Running: testSelectAcrossMultipleMessagesWithTruncatedUrls

Making selection: 0, 0, 2, 75
Selection length: 260, text: www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com
VM:5 www.bar.com
VM:6 www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z

Making selection: 0, 0, 2, 76
Selection length: 269, text: www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com
VM:5 www.bar.com
VM:6 www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789

Making selection: 0, 0, 2, 150
Selection length: 343, text: www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com
VM:5 www.bar.com
VM:6 www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com

Running: testSelectWithinMessageWithMultipleTruncatedUrls

Making selection: 3, 0, 3, 75
Selection length: 75, text: www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z

Making selection: 3, 0, 3, 76
Selection length: 84, text: www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789

Making selection: 3, 0, 3, 163
Selection length: 171, text: www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com www.bar.com 

Making selection: 3, 0, 3, 238
Selection length: 246, text: www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com www.bar.com www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z

Making selection: 3, 0, 3, 239
Selection length: 255, text: www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com www.bar.com www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789

Making selection: 3, 0, 3, 313
Selection length: 329, text: www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com www.bar.com www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com

Making selection: 3, 75, 3, 76
Selection length: 9, text: 123456789

Making selection: 3, 75, 3, 163
Selection length: 96, text: 123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com www.bar.com 

Making selection: 3, 75, 3, 238
Selection length: 171, text: 123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com www.bar.com www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z

Making selection: 3, 75, 3, 239
Selection length: 180, text: 123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com www.bar.com www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789

Making selection: 3, 75, 3, 313
Selection length: 254, text: 123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com www.bar.com www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com

Making selection: 3, 76, 3, 163
Selection length: 87, text: z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com www.bar.com 

Making selection: 3, 76, 3, 238
Selection length: 162, text: z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com www.bar.com www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z

Making selection: 3, 76, 3, 239
Selection length: 171, text: z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com www.bar.com www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789

Making selection: 3, 76, 3, 313
Selection length: 245, text: z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com www.bar.com www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com

Making selection: 3, 163, 3, 238
Selection length: 75, text: www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z

Making selection: 3, 163, 3, 239
Selection length: 84, text: www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789

Making selection: 3, 163, 3, 313
Selection length: 158, text: www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com

Making selection: 3, 238, 3, 239
Selection length: 9, text: 123456789

Making selection: 3, 238, 3, 313
Selection length: 83, text: 123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com

Making selection: 3, 239, 3, 313
Selection length: 74, text: z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com

Running: testSelectWithinShortUrlWithHashes

Making selection: 4, 0, 4, 14
Selection length: 26, text: www.01234567890123456789zf

Making selection: 4, 0, 4, 28
Selection length: 52, text: www.01234567890123456789zfoobarz01234567890123456789

Making selection: 4, 14, 4, 28
Selection length: 26, text: oobarz01234567890123456789

Running: testSelectWithinUrlWithHashes

Making selection: 5, 0, 5, 63
Selection length: 75, text: www.01234567890123456789zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz

Making selection: 5, 0, 5, 64
Selection length: 120, text: www.01234567890123456789zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz

Making selection: 5, 0, 5, 126
Selection length: 194, text: www.01234567890123456789zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz01234567890123456789

Making selection: 5, 63, 5, 64
Selection length: 45, text: zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz

Making selection: 5, 63, 5, 126
Selection length: 119, text: zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz01234567890123456789

Making selection: 5, 64, 5, 126
Selection length: 74, text: zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz01234567890123456789

Running: testSelectWithinHighlightedUrlBeginning
Searching for text: www.
Highlighted 10 matches

Making selection: 6, 0, 6, 75
Selection length: 75, text: www.zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz

Making selection: 6, 0, 6, 76
Selection length: 134, text: www.zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz

Making selection: 6, 0, 6, 150
Selection length: 208, text: www.zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz.com

Making selection: 6, 75, 6, 76
Selection length: 59, text: zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz

Making selection: 6, 75, 6, 150
Selection length: 133, text: zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz.com

Making selection: 6, 76, 6, 150
Selection length: 74, text: zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz.com

Running: testSelectWithinHighlightedUrlMiddle
Searching for text: zzzzz
Highlighted 48 matches

Making selection: 6, 0, 6, 75
Selection length: 75, text: www.zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz

Making selection: 6, 0, 6, 76
Selection length: 134, text: www.zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz

Making selection: 6, 0, 6, 150
Selection length: 208, text: www.zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz.com

Making selection: 6, 75, 6, 76
Selection length: 59, text: zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz

Making selection: 6, 75, 6, 150
Selection length: 133, text: zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz.com

Making selection: 6, 76, 6, 150
Selection length: 74, text: zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz.com

Running: testSelectWithinHighlightedUrlEnd
Searching for text: .com
Highlighted 8 matches

Making selection: 6, 0, 6, 75
Selection length: 75, text: www.zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz

Making selection: 6, 0, 6, 76
Selection length: 134, text: www.zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz

Making selection: 6, 0, 6, 150
Selection length: 208, text: www.zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz.com

Making selection: 6, 75, 6, 76
Selection length: 59, text: zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz

Making selection: 6, 75, 6, 150
Selection length: 133, text: zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz.com

Making selection: 6, 76, 6, 150
Selection length: 74, text: zzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzzz.com

Running: testSelectionStartingInNonTextNode

Making selection from empty prefix to end
Selection length: 166, text: VM:11 www.z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789z123456789.com
